[refactor] Sign 부분 style component 분리 및 컴포넌트 분리
